// src/app/models/Tailor.ts
export interface Tailor {
  id?: string;
  tailorName?: string;
  shopName?: string;
  shopNumber?: string;
  description?: string;
  dateOfRegistration?: Date;
  street?: string;
  area?: string;
  city?: string;
  state?: string;
  country?: string;
  mobile?: string;
  email?: string;
  maleDress?: string;
  femaleDress?: string;
  tags?: string;
  imageUrl?: string;
}